- LocationManager update to use only FusedLocation - Samples: show list ordered by name
