+++ sample/src/main/java/com/massivedisaster/adal/sample/feature/home/FragmentHome.java
@@ -32,13 +32,14 @@ import com.massivedisaster.adal.sample.base.activity.ActivityToolbar;
 import com.massivedisaster.adal.sample.feature.accounts.FragmentAccounts;
 import com.massivedisaster.adal.sample.feature.analytics.FragmentAnalytics;
 import com.massivedisaster.adal.sample.feature.bus.FragmentA;
-import com.massivedisaster.adal.sample.feature.location.FragmentFusedLocation;
 import com.massivedisaster.adal.sample.feature.location.FragmentLocation;
 import com.massivedisaster.adal.sample.feature.network.FragmentNetworkRequest;
 import com.massivedisaster.adal.sample.feature.permissions.FragmentPermissions;
 import com.massivedisaster.adal.sample.feature.utils.FragmentUtils;
 
 import java.util.HashMap;
+import java.util.SortedMap;
+import java.util.TreeMap;
 
 public class FragmentHome extends AbstractBaseFragment {
 
@@ -79,9 +80,8 @@ public class FragmentHome extends AbstractBaseFragment {
         mRclItems.setAdapter(adapter);
     }
 
-    public HashMap<String, Class<? extends Fragment>> getExamples() {
-        return new HashMap<String, Class<? extends Fragment>>() {{
-            put("FusedLocation", FragmentFusedLocation.class);
+    public TreeMap<String, Class<? extends Fragment>> getExamples() {
+        return new TreeMap<String, Class<? extends Fragment>>() {{
             put("Location", FragmentLocation.class);
             put("Permission", FragmentPermissions.class);
             put("Accounts", FragmentAccounts.class);
